Merge with mainline to get latest APIs and device tree bindings.

+++ include/media/videobuf-dma-sg.h
@@ -61,7 +61,7 @@ struct videobuf_dmabuf {
 	/* common */
 	struct scatterlist  *sglist;
 	int                 sglen;
-	int                 nr_pages;
+	unsigned long       nr_pages;
 	int                 direction;
 };
